+++ src/pl/testuj/operations/CAddOperation.java
@@ -4,7 +4,7 @@ package pl.testuj.operations;
  * Created by krzysztoffonal on 10/09/16.
  */
 public class CAddOperation extends COperation {
-  @Override public double process(double firstArg, double secondArg) {
+  @Override public int process(int firstArg, int secondArg) {
     return firstArg + secondArg;
   }
 }
