+++ src/pl/testuj/operations/COperation.java
@@ -4,5 +4,5 @@ package pl.testuj.operations;
  * Created by krzysztoffonal on 10/09/16.
  */
 public abstract class COperation {
-  public abstract double process(double firstArg, double secondArg);
+  public abstract int process(int firstArg, int secondArg);
 }
